% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscis-package.R
\docType{package}
\name{uscis-package}
\alias{uscis}
\alias{uscis-package}
\title{uscis: USCIS Immigration Form Processing Times}
\description{
Provides functions to access and analyze USCIS (United States Citizenship and Immigration Services) processing times data through their API. Includes functionality to retrieve form information, office locations, and current processing times for various immigration forms.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coatless-r-n-d/uscis-processing}
  \item \url{https://rd.thecoatlessprofessor.com/uscis-processing/}
}

}
\author{
\strong{Maintainer}: James Joseph Balamuta \email{balamut2@illinois.edu}

}
\keyword{internal}
