% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{uscis_form_offices}
\alias{uscis_form_offices}
\title{USCIS Form Offices}
\usage{
uscis_form_offices(form_types, cookie_file = "uscis.cookies")
}
\arguments{
\item{form_types}{Data frame from \code{uscis_form_types()}}

\item{cookie_file}{Path to cookie file}
}
\value{
Data frame of forms and their processing offices with the following columns:
\describe{
\item{form_name}{Character. The name/identifier of the USCIS form}
\item{form_type}{Character. Type classification of the form}
\item{office_code}{Character. The USCIS office location code}
\item{office_description}{Character. Description of the USCIS Office Code in English}
}
}
\description{
Retrieves processing offices for each form type
}
\examples{
\dontrun{
# Initialize USCIS session
uscis_init_session()

# Retrieve USCIS forms
forms <- uscis_forms()

# Retrieve form types based on form names
form_types <- uscis_form_types(forms)

# Retrieve processing offices for each form type
form_offices <- uscis_form_offices(form_types)
head(form_offices)
}
}
