% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{uscis_form_types}
\alias{uscis_form_types}
\title{USCIS Form Types}
\usage{
uscis_form_types(forms, cookie_file = "uscis.cookies")
}
\arguments{
\item{forms}{Data frame of forms from \code{uscis_forms()}}

\item{cookie_file}{Path to cookie file}
}
\value{
Data frame of form types and subtypes with the following columns:
\describe{
\item{form_name}{Character. The name/identifier of the USCIS form}
\item{form_key}{Character. Unique identifier key for the form}
\item{form_type}{Character. Type classification of the form}
\item{form_type_description_en}{Character. Detailed description of the form type in English}
\item{form_type_description_es}{Character. Detailed description of the form type in Spanish}
}
}
\description{
Retrieves available subtypes for each USCIS form
}
\examples{
\dontrun{
# Initialize USCIS session
uscis_init_session()

# Retrieve USCIS forms
forms <- uscis_forms()

# Retrieve form types based on form names
form_types <- uscis_form_types(forms)
head(form_types)
}
}
