% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{uscis_forms}
\alias{uscis_forms}
\title{USCIS Forms}
\usage{
uscis_forms(cookie_file = "uscis.cookies")
}
\arguments{
\item{cookie_file}{Path to cookie file}
}
\value{
A data frame with the following columns:
\describe{
\item{form_name}{Character. The name/identifier of the USCIS form}
\item{form_description_en}{Character. Description of the USCIS form in English}
\item{form_description_es}{Character. Description of the USCIS form in Spanish}
}
}
\description{
Retrieves list of available USCIS forms
}
\examples{
\dontrun{
# Initialize USCIS session
uscis_init_session()

# Retrieve USCIS forms
forms <- uscis_forms()
head(forms)
}
}
