% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{uscis_init_session}
\alias{uscis_init_session}
\title{Initialize USCIS API Session}
\usage{
uscis_init_session(cookie_file = "uscis.cookies")
}
\arguments{
\item{cookie_file}{Path to save cookies. Default is "uscis.cookies" in the working directory.}
}
\value{
Boolean if session was successfully initialized
}
\description{
Makes initial request to establish session cookies
}
