% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{uscis_query}
\alias{uscis_query}
\title{Make USCIS API Request}
\usage{
uscis_query(endpoint, cookie_file = "uscis.cookies")
}
\arguments{
\item{endpoint}{API endpoint to query}

\item{cookie_file}{Path to cookie file. Default is "uscis.cookies" in the working directory.}
}
\value{
Parsed JSON response
}
\description{
Internal function to make authenticated requests to USCIS API
}
\keyword{internal}
